Moved 4 functions to FirebaseManager
