+++ Android/app/src/main/java/hr/foi/air/fitfusion/ReplyActivity.kt
@@ -3,7 +3,6 @@ package hr.foi.air.fitfusion
 import android.annotation.SuppressLint
 import android.app.AlertDialog
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.widget.Button
 import android.widget.EditText
@@ -11,14 +10,9 @@ import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.google.firebase.database.DataSnapshot
-import com.google.firebase.database.DatabaseError
-import com.google.firebase.database.FirebaseDatabase
-import com.google.firebase.database.ValueEventListener
 import hr.foi.air.fitfusion.adapters.ReplyAdapter
 import hr.foi.air.fitfusion.data_classes.FirebaseManager
 import hr.foi.air.fitfusion.data_classes.LoggedInUser
-import hr.foi.air.fitfusion.entities.Reply
 import java.text.SimpleDateFormat
 import java.util.Calendar
 import java.util.Locale
@@ -28,8 +22,6 @@ class ReplyActivity : AppCompatActivity(){
     private lateinit var recyclerView: RecyclerView
     private lateinit var idPost: String
     private lateinit var firebaseManager: FirebaseManager
-    private val database = FirebaseDatabase.getInstance()
-    private val postRef = database.getReference("Replies")
     private lateinit var loggedInUser: LoggedInUser
     private lateinit var replyAdapter: ReplyAdapter
     @SuppressLint("SetTextI18n")
@@ -67,7 +59,7 @@ class ReplyActivity : AppCompatActivity(){
         btnPostReply.setOnClickListener {
             showDialog()
         }
-        fetchReplyFromFirebase(postTitle)
+        firebaseManager.fetchReplyFromFirebase(postTitle, recyclerView)
     }
 
     private fun convertTimestampToDateString(timestamp: Long): String {
@@ -97,66 +89,9 @@ class ReplyActivity : AppCompatActivity(){
                 val authorFirstName = loggedInUser.getFirstName()
                 val authorLastName = loggedInUser.getLastName()
                 val timestamp = System.currentTimeMillis()
-                saveReplyToFirebase(content, timestamp, authorFirstName, authorLastName)
+                firebaseManager.saveReplyToFirebase(content, timestamp, authorFirstName, authorLastName, idPost)
                 dialog.dismiss()
             }
-        }
-    }
-    private fun saveReplyToFirebase(content: String, timestamp: Long, authorFirstName: String?, authorLastName: String?){
-        val replyId = postRef.push().key
-        val newPost =
-            mapOf(
-                "id" to replyId,
-                "content" to content,
-                "timestamp" to timestamp,
-                "authorFirstName" to authorFirstName,
-                "authorLastName" to authorLastName,
-                "postId" to idPost
-            )
-
-        if (replyId != null){
-            postRef.child(replyId).setValue(newPost)
-                .addOnSuccessListener {}
-                .addOnFailureListener {}
-        }
-    }
-
-    private fun fetchReplyFromFirebase(postTitle: String) {
-
-        firebaseManager.getPostId(postTitle) { postId ->
-            val query = postRef.orderByChild("postId").equalTo(postId)
-
-
-            query.addValueEventListener(object : ValueEventListener {
-                override fun onDataChange(dataSnapshot: DataSnapshot) {
-                    val replies: MutableList<Reply> = mutableListOf()
-
-                    for (postSnapshot in dataSnapshot.children) {
-                        val id = postSnapshot.child("id").getValue(String::class.java) ?: ""
-                        val postIds = postSnapshot.child("postId").getValue(String::class.java) ?: ""
-                        val content =
-                            postSnapshot.child("content").getValue(String::class.java) ?: ""
-                        val timestamp =
-                            postSnapshot.child("timestamp").getValue(Long::class.java) ?: 0
-                        val authorFirstName =
-                            postSnapshot.child("authorFirstName").getValue(String::class.java) ?: ""
-                        val authorLastName =
-                            postSnapshot.child("authorLastName").getValue(String::class.java) ?: ""
-
-
-                        val reply =
-                            Reply(id, postIds, content, timestamp, authorFirstName, authorLastName)
-                        replies.add(reply)
-                    }
-
-                    replyAdapter = ReplyAdapter(replies)
-                    recyclerView.adapter = replyAdapter
-                }
-
-                override fun onCancelled(databaseError: DatabaseError) {
-                    Log.d("ReplyActivity", "onCancelled: ${databaseError.message}")
-                }
-            })
         }
     }
 }
